move self-service-role logic into own module
