+++ reconcile/test/test_change_owners.py
@@ -6,9 +6,11 @@ from reconcile.changeowners.diff import (
     deepdiff_path_to_jsonpath,
 )
 from reconcile.change_owners import (
-    cover_changes_with_self_service_roles,
     apply_decisions_to_changes,
     manage_conditional_label,
+)
+from reconcile.changeowners.selfserviceroles import (
+    cover_changes_with_self_service_roles
 )
 from reconcile.changeowners.decision import (
     get_approver_decisions_from_mr_comments,
